LPS-188165 Extract logic for ItemSelectorViewDescriptor into a new class, no logic changes
